requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
deepface==0.0.93
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
numpy==1.26.4
onnx==1.17.0
onnxconverter-common==1.14.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
skl2onnx==1.18.0
soupsieve==2.6
SQLAlchemy==2.0.38
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
